+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog (niondata)
 UNRELEASED
 ----------
 
+- Allow persistent relationships to define storage key.
+
 - Extend Registry to allow registering same component with additional component types.
 
 0.3.14 (2018-09-13)